arm: zynq: Rework green-mango.dts
The hardware design changed dramatically recently. A
lot of new features were added. There is still some
work to be done on the .dts file so that all the
hardware features are represented.

This is a work-in-progress change. Note the commented
out device tree nodes. Said nodes need to be tweaked
and uncommented in a later commit.