arm: green-mango: Fix gpio-keys trigger
This applies to both the Zeus and BactoBox board.

Previously, the gpio-keys were triggered by a gpios entry. In practice,
this caused the key release event to emit before the key press event.

Now, the gpio-keys are triggered by a interrupts entry. This ensures
the correct key press/release sequence.