arm: green-mango: Fix IRQ spam during a key press on the Zeus board
Changed interrupt type from LEVEL_LOW to EDGE_FALLING. With LEVEL_LOW, the kernel
was spammed with interrupts while the button was held down. This could even
cause a kernel crash.

Now EDGE_FALLING is used instead, which matches the hardware configuration.
The newest version of the FPGA image fixes a bug with the GPIO keys due to
an erroneous pullup. In turn, this enables the use of EDGE_FALLING.