arm: green-mango: fix incorrect use of gpio0
The gpio descriptor only has 2 cells but we supplied 3.
In practice, the 3rd cell was silently discarded. The main problem
was that the 2nd cell was thought to be an 'offset' parameter that
always had the value of 0. In practice, however, the 2nd cell was
the 'flags' parameter and a value of 0 translates to GPIO_ACTIVE_HIGH.

Consequently, whenever we thought that we set GPIO_ACTIVE_LOW for
a gpio0-based GPIO, the actual flags remained at GPIO_ACTIVE_HIGH.

The gpio-keys used interrupts directly because we saw strange behaviour
with the GPIO-based approach. Now that this issue is solved, the
strange behaviour suddenly makes sense.