arm: green-mango: rebase defconfig
Previously, our bactobox and zeus defconfigs were based on the Zed board.
Now, we base them on the new xilinx_zynq_virt_defconfig (on which the Zed
board is also based now).

Also, we are nearing the point where we can have a single green-mango
defconfig for both the bactobox and zeus boards.